+++ script/pload.ps1
@@ -1,2 +1,2 @@
 # Compatibility file for users who prefer dot-sourcing a script directly.
-Invoke-Expression (& python_virtual_env_load shell-init powershell | Out-String)
+Invoke-Expression (& pload shell-init powershell | Out-String)