+++ script/pload.zsh
@@ -1,2 +1,2 @@
 # Compatibility file for users who prefer sourcing a script directly.
-eval "$(python_virtual_env_load shell-init zsh)"
+eval "$(pload shell-init zsh)"